pyproject.toml
[project]
name = "sdofmv2"
version = "0.1.0"
description = "Solar phenomena prediction models"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Joseph Gallego",     email = "joaggi@gmail.com"},
    {name = "Daniela Martin",     email = "dmartinvega@gmail.com"},
    {name = "Jinsu Hong",         email = "jinsuhong.knight@gmail.com"},
]
keywords = ["foundation model", "solar physics", "deep learning", "space weather"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Astronomy",
]

dependencies = [
    "torch>=2.5.1",
    "lightning>=2.6.0",
    "numpy>=2.3.5",
    "pandas>=2.3.3",
    "transformers>=4.57.3",
    "sunpy>=7.0.4",
    "astropy>=6.0",
    "timm>=1.0.22",
    "einops>=0.8.1",
    "hydra-core>=1.3.2",
    "wandb>=0.23.1",
]

[project.urls]
Repository = "https://github.com/Joaggi/sdofmv2"
Issues = "https://github.com/Joaggi/sdofmv2/issues"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
